+++ crates/fuel-core/src/service.rs
@@ -333,6 +333,14 @@ impl FuelService {
             }
         }
 
+        // repopulate missing tables
+        genesis::recover_missing_tables_from_genesis_state_config(
+            watcher.clone(),
+            &self.shared.config,
+            &self.shared.database,
+        )
+        .await?;
+
         self.override_chain_config_if_needed()
     }
 }